Split config into multiple files
This file was getting big, so I think it's about time to start splitting
the module up.
