--- src/config.rs
@@ -1,79 +1,0 @@
-use config::ConfigError;
-
-use serde_aux::field_attributes::deserialize_number_from_string;
-
-#[derive(serde::Deserialize, Clone, Debug)]
-pub struct Settings {
-    pub application: ApplicationSettings,
-    pub database: DatabaseSettings,
-    pub admin: AdminSettings,
-}
-
-#[derive(serde::Deserialize, Clone, Debug, Default)]
-pub struct ApplicationSettings {
-    #[serde(deserialize_with = "deserialize_number_from_string")]
-    pub port: u16,
-    pub host: String,
-}
-
-#[derive(serde::Deserialize, Clone, Debug, Default)]
-pub struct DatabaseSettings {
-    pub url: String,
-}
-
-#[derive(serde::Deserialize, Clone, Debug, Default)]
-pub struct AdminSettings {
-    pub token: String,
-}
-
-pub fn get_configuration() -> Result<Settings, ConfigError> {
-    let base_path = std::env::current_dir().expect("Failed to determine the current directory");
-    let configuration_directory = base_path.join("config");
-
-    let environment: Environment = std::env::var("ENVIRONMENT")
-        .unwrap_or_else(|_| "local".into())
-        .try_into()
-        .expect("Failed to parse ENVIRONMENT.");
-
-    let environment_filename = format!("{}.yaml", environment.as_str());
-
-    let settings = config::Config::builder()
-        .add_source(config::File::from(
-            configuration_directory.join("base.yaml"),
-        ))
-        .add_source(config::File::from(
-            configuration_directory.join(environment_filename),
-        ))
-        .add_source(config::Environment::default().separator("_"))
-        .build()?;
-
-    settings.try_deserialize()
-}
-
-pub enum Environment {
-    Local,
-    Production,
-}
-
-impl Environment {
-    pub fn as_str(&self) -> &'static str {
-        match self {
-            Environment::Local => "local",
-            Environment::Production => "production",
-        }
-    }
-}
-
-impl TryFrom<String> for Environment {
-    type Error = String;
-
-    fn try_from(s: String) -> Result<Self, Self::Error> {
-        match s.to_lowercase().as_str() {
-            "local" => Ok(Self::Local),
-            "production" => Ok(Self::Production),
-            other => Err(format!(
-                "{other} is not a supported environment. Use either `local` or `production`."
-            )),
-        }
-    }
-}
