+++ src/config/admin_settings.rs
@@ -1,0 +1,4 @@
+#[derive(serde::Deserialize, Clone, Debug, Default)]
+pub struct AdminSettings {
+    pub token: String,
+}
