+++ src/config/database_settings.rs
@@ -1,0 +1,4 @@
+#[derive(serde::Deserialize, Clone, Debug, Default)]
+pub struct DatabaseSettings {
+    pub url: String,
+}
